organize librarys move nlp to data-scrape (requires requests)
