+++ src/data-scratch/data_science_from_scratch/clustering.py
@@ -3,7 +3,7 @@ import random
 from logging.config import dictConfig
 
 from data_science_from_scratch import config
-from data_science_from_scratch.linear_algebra import (
+from data_science_from_scratch.library.linear_algebra import (
     squared_distance,
     vector_mean,
     distance,
