+++ src/data-scratch/data_science_from_scratch/library/__init__.py
@@ -1,0 +1,3 @@
+import logging
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())
