+++ src/data-scratch/data_science_from_scratch/library/linear_algebra.py
