+++ src/data-scratch/data_science_from_scratch/logistic_regression.py
@@ -5,12 +5,12 @@ from functools import partial, reduce
 from logging.config import dictConfig
 
 from data_science_from_scratch import config
-from data_science_from_scratch.gradient_descent import (
+from data_science_from_scratch.library.gradient_descent import (
     maximize_stochastic,
     maximize_batch,
 )
-from data_science_from_scratch.linear_algebra import dot, vector_add
-from data_science_from_scratch.machine_learning import train_test_split
+from data_science_from_scratch.library.linear_algebra import dot, vector_add
+from data_science_from_scratch.library.machine_learning import train_test_split
 from data_science_from_scratch.multiple_regression import estimate_beta
 from data_science_from_scratch.working_with_data import rescale
