+++ src/data-scratch/data_science_from_scratch/recommender_systems.py
@@ -4,7 +4,7 @@ from collections import defaultdict, Counter
 from logging.config import dictConfig
 
 from data_science_from_scratch import config
-from data_science_from_scratch.linear_algebra import dot
+from data_science_from_scratch.library.linear_algebra import dot
 
 users_interests = [
     ["Hadoop", "Big Data", "HBase", "Java", "Spark", "Storm", "Cassandra"],
